majority element
C program that takes a positive integer n and an array A of n integers as input and does the following. If there is an element which appears more than n/2 times in the array A, the program prints that element as output. Otherwise, the program prints "No Majority". Make sure the number of comparisons of your program is at most quadratic with respect to n and the additional space used is bounded by a constant, independent of the input.